Address Alan's comments
